add config/, bookingplanservice_db+ planservice_db + user_handle

+++ model/field.go
@@ -1,7 +1,7 @@
 package model
 
 type Field struct {
-	ID   uint   `gorm:"primaryKey;auto_increment"`
+	ID   uint64 `gorm:"primaryKey;auto_increment"`
 	Name string `gorm:"not null;unique"`
 }
